+++ docker_compose_driver.go
@@ -72,12 +72,14 @@ func (dc DockerComposeDriver) getDCGeneratedFilePath(dcfilePath string) string {
 	return dcfilePath + ".dojo"
 }
 
-func (dc DockerComposeDriver) generateDCFileContentsWithEnv(expContainers []string,	config Config, envFile string) string {
+func (dc DockerComposeDriver) generateDCFileContentsWithEnv(expContainers []string,	config Config, envFile string,
+	envFileMultiLine string) string {
 	contents := fmt.Sprintf(
 		`    volumes:
       - %s:%s:ro
       - %s:%s
-`, config.IdentityDirOuter, "/dojo/identity", config.WorkDirOuter, config.WorkDirInner)
+      - %s:/etc/dojo.d/variables/00-multiline-vars.sh
+`, config.IdentityDirOuter, "/dojo/identity", config.WorkDirOuter, config.WorkDirInner, envFileMultiLine)
 	if os.Getenv("DISPLAY") != "" {
 		// DISPLAY is set, enable running in graphical mode (opinionated)
 		contents += "      - /tmp/.X11-unix:/tmp/.X11-unix\n"
@@ -96,7 +98,9 @@ func (dc DockerComposeDriver) generateDCFileContentsWithEnv(expContainers []stri
 			contents += fmt.Sprintf(`  %s:
     env_file:
       - %s
-`, name, envFile)
+    volumes:
+      - %s:/etc/dojo.d/variables/00-multiline-vars.sh
+`, name, envFile, envFileMultiLine)
 		}
 	}
 	return contents
@@ -320,14 +324,14 @@ func safelyCloseChannel(ch chan bool) (justClosed bool) {
 
 func (dc DockerComposeDriver) HandleRun(mergedConfig Config, runID string, envService EnvServiceInterface) int {
 	warnGeneral(dc.FileService, mergedConfig, envService, dc.Logger)
-	envFile := getEnvFilePath(runID, mergedConfig.Test)
-	saveEnvToFile(dc.FileService, envFile, mergedConfig.BlacklistVariables, envService.GetVariables())
+	envFile, envFileMultiLine := getEnvFilePaths(runID, mergedConfig.Test)
+	saveEnvToFile(dc.FileService, envFile, envFileMultiLine, mergedConfig.BlacklistVariables, envService.GetVariables())
 	dojoDCGeneratedFile, err := dc.handleDCFiles(mergedConfig)
 	if err != nil {
 		return 1
 	}
 	expContainers := dc.getExpectedContainers(mergedConfig, runID)
-	additionalContents := dc.generateDCFileContentsWithEnv(expContainers, mergedConfig, envFile)
+	additionalContents := dc.generateDCFileContentsWithEnv(expContainers, mergedConfig, envFile, envFileMultiLine)
 	dc.FileService.AppendContents(dojoDCGeneratedFile, additionalContents, "debug")
 
 	cmd := dc.ConstructDockerComposeCommandRun(mergedConfig, runID)
@@ -354,8 +358,9 @@ func (dc DockerComposeDriver) HandleRun(mergedConfig Config, runID string, envSe
 func (dc DockerComposeDriver) CleanAfterRun(mergedConfig Config, runID string) int {
 	if mergedConfig.RemoveContainers == "true" {
 		dc.Logger.Log("debug", "Cleaning, because RemoveContainers is set to true")
-		envFile := getEnvFilePath(runID, mergedConfig.Test)
+		envFile, envFileMultiLine := getEnvFilePaths(runID, mergedConfig.Test)
 		defer dc.FileService.RemoveGeneratedFile(mergedConfig.RemoveContainers, envFile)
+		defer dc.FileService.RemoveGeneratedFile(mergedConfig.RemoveContainers, envFileMultiLine)
 		dojoDCGeneratedFile := dc.getDCGeneratedFilePath(mergedConfig.DockerComposeFile)
 		defer dc.FileService.RemoveGeneratedFile(mergedConfig.RemoveContainers, dojoDCGeneratedFile)