+++ docker_compose_driver_test.go
@@ -95,34 +95,45 @@ func Test_generateDCFileContentsWithEnv(t *testing.T){
 		} else {
 			setTestEnv()
 		}
-		contents := dc.generateDCFileContentsWithEnv(expectedServices, config, "/tmp/env-file.txt")
+		contents := dc.generateDCFileContentsWithEnv(expectedServices, config,
+			"/tmp/env-file.txt", "/tmp/env-file-multiline.txt")
 
 		if v.displaySet {
 			assert.Equal(t,  `    volumes:
       - /tmp/myidentity:/dojo/identity:ro
       - /tmp/bla:/dojo/work
+      - /tmp/env-file-multiline.txt:/etc/dojo.d/variables/00-multiline-vars.sh
       - /tmp/.X11-unix:/tmp/.X11-unix
     env_file:
       - /tmp/env-file.txt
   abc:
     env_file:
       - /tmp/env-file.txt
+    volumes:
+      - /tmp/env-file-multiline.txt:/etc/dojo.d/variables/00-multiline-vars.sh
   def:
     env_file:
       - /tmp/env-file.txt
+    volumes:
+      - /tmp/env-file-multiline.txt:/etc/dojo.d/variables/00-multiline-vars.sh
 `, contents)
 		} else {
 			assert.Equal(t,  `    volumes:
       - /tmp/myidentity:/dojo/identity:ro
       - /tmp/bla:/dojo/work
+      - /tmp/env-file-multiline.txt:/etc/dojo.d/variables/00-multiline-vars.sh
     env_file:
       - /tmp/env-file.txt
   abc:
     env_file:
       - /tmp/env-file.txt
+    volumes:
+      - /tmp/env-file-multiline.txt:/etc/dojo.d/variables/00-multiline-vars.sh
   def:
     env_file:
       - /tmp/env-file.txt
+    volumes:
+      - /tmp/env-file-multiline.txt:/etc/dojo.d/variables/00-multiline-vars.sh
 `, contents)
 		}
 	}
@@ -286,18 +297,23 @@ func TestDockerComposeDriver_HandleRun_Unit(t *testing.T) {
 	config.Driver = "docker-compose"
 	config.RunCommand = "bla"
 	runID := "1234"
-	exitstatus := driver.HandleRun(config, runID, NewMockedEnvService())
+	envService := NewMockedEnvService()
+	envService.AddVariable(`MULTI_LINE=one
+two
+three`)
+	exitstatus := driver.HandleRun(config, runID, envService)
 	assert.Equal(t, 0, exitstatus)
-	assert.Equal(t, 2, len(fs.FilesWrittenTo))
+	assert.Equal(t, 3, len(fs.FilesWrittenTo))
 	assert.Equal(t, "ABC=123\n", fs.FilesWrittenTo["/tmp/dojo-environment-1234"])
+	assert.Equal(t, "export MULTI_LINE=$(echo b25lCnR3bwp0aHJlZQ== | base64 -d)\n", fs.FilesWrittenTo["/tmp/dojo-environment-multiline-1234"])
 	assert.Contains(t, fs.FilesWrittenTo["docker-compose.yml.dojo"], "version: '2.2'")
 
 	exitstatus = driver.CleanAfterRun(config, runID)
 	assert.Equal(t, 0, exitstatus)
-	assert.Equal(t, 3, len(fs.FilesRemovals))
+	assert.Equal(t, 5, len(fs.FilesRemovals))
 	assert.Equal(t, "/tmp/dojo-environment-1234", fs.FilesRemovals[0])
-	assert.Equal(t, "docker-compose.yml.dojo", fs.FilesRemovals[1])
-	assert.Equal(t, "/tmp/dojo-environment-1234", fs.FilesRemovals[2])
+	assert.Equal(t, "/tmp/dojo-environment-multiline-1234", fs.FilesRemovals[1])
+	assert.Equal(t, "docker-compose.yml.dojo", fs.FilesRemovals[2])
 	assert.False(t, fileExists("/tmp/dojo-environment-1234"))
 }